# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/winning-pattern": {
      "post": {
        "operationId": "WinningPatternController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWinningPatternDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      },
      "get": {
        "operationId": "WinningPatternController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      }
    },
    "/winning-pattern/{id}": {
      "get": {
        "operationId": "WinningPatternController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      },
      "patch": {
        "operationId": "WinningPatternController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWinningPatternDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      },
      "delete": {
        "operationId": "WinningPatternController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateWinningPatternDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Line"
          },
          "slug": {
            "type": "string",
            "description": "Slug (auto-generated from name if not provided)",
            "example": "line"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "pattern_type": {
            "type": "string",
            "description": "Pattern type: BASIC (all users) or ADVANCED (premium only)",
            "example": "BASIC",
            "enum": [
              "BASIC",
              "ADVANCED"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateWinningPatternDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Line"
          },
          "slug": {
            "type": "string",
            "description": "Slug",
            "example": "line"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "pattern_type": {
            "type": "string",
            "description": "Pattern type: BASIC or ADVANCED",
            "enum": [
              "BASIC",
              "ADVANCED"
            ]
          }
        }
      },
...
    }
  }
}